temp/py/requirements.txt
Django==4.1.5
django-cors-headers
nltk==3.8.1
sentence-transformers
requests
beautifulsoup4
torch
torchvision
scikit-learn
numpy
transformers
djangorestframework